+++ src/main/java/net/jptrzy/trinkets/curios/theme/config/ModConfig.java
@@ -3,4 +3,9 @@ package net.jptrzy.trinkets.curios.theme.config;
 public class ModConfig {
     public static boolean scrollbar = true;
     public static boolean scrolling_outside_boundary = false;
+
+    public static int max_height = 7;
+    public static int min_width = 1;
+
+    public static boolean scout_auto_resize = false;
 }