+++ src/main/java/net/jptrzy/trinkets/curios/theme/config/ModConfigData.java
@@ -7,9 +7,16 @@ import net.jptrzy.trinkets.curios.theme.Client;
 
 @Config(name = Client.MOD_ID)
 public class ModConfigData implements ConfigData {
-//    @ConfigEntry.Category("Client")
     public boolean scrollbar = true;
-    @ConfigEntry.Gui.Tooltip
-    public boolean scrolling_outside_boundary = false;
+    @ConfigEntry.Gui.PrefixText public boolean scrolling_outside_boundary = false;
 
+    @ConfigEntry.BoundedDiscrete(min = 2, max = 7)
+    public int max_height = 7;
+    @ConfigEntry.Gui.PrefixText
+    @ConfigEntry.BoundedDiscrete(min = 1, max = 4)
+    public int min_width = 1;
+
+//    @ConfigEntry.Category("Integrations")
+    // TODO make other inputs gray color when scout detected and this option set to true
+    @ConfigEntry.Gui.PrefixText public boolean scout_auto_resize = true;
 }